Modernize the test suite
Instead of using so much custom homegrown, we'll instead use pytest as our test
runner and to generate test cases. We'll also do some logic clean up in the
tests and add tox to make it easier to run all of the tests.
