+++ tests/data/invalid-python-name-module-2.loader-error
@@ -1,1 +1,1 @@
---- !!python/name:xml.parsers
+--- !!python/name:xml.nonexistent
