+++ tests/test_errors.py
@@ -1,12 +1,14 @@
 
-import yaml, test_emitter
+import yaml
+
+from tests.test_emitter import EventsLoader
 
 def test_loader_error(error_filename, verbose=False):
     try:
         list(yaml.load_all(open(error_filename, 'rb')))
-    except yaml.YAMLError, exc:
+    except yaml.YAMLError as exc:
         if verbose:
-            print "%s:" % exc.__class__.__name__, exc
+            print("%s:" % exc.__class__.__name__, exc)
     else:
         raise AssertionError("expected an exception")
 
@@ -15,9 +17,9 @@ test_loader_error.unittest = ['.loader-error']
 def test_loader_error_string(error_filename, verbose=False):
     try:
         list(yaml.load_all(open(error_filename, 'rb').read()))
-    except yaml.YAMLError, exc:
+    except yaml.YAMLError as exc:
         if verbose:
-            print "%s:" % exc.__class__.__name__, exc
+            print("%s:" % exc.__class__.__name__, exc)
     else:
         raise AssertionError("expected an exception")
 
@@ -26,9 +28,9 @@ test_loader_error_string.unittest = ['.loader-error']
 def test_loader_error_single(error_filename, verbose=False):
     try:
         yaml.load(open(error_filename, 'rb').read())
-    except yaml.YAMLError, exc:
+    except yaml.YAMLError as exc:
         if verbose:
-            print "%s:" % exc.__class__.__name__, exc
+            print("%s:" % exc.__class__.__name__, exc)
     else:
         raise AssertionError("expected an exception")
 
@@ -36,12 +38,12 @@ test_loader_error_single.unittest = ['.single-loader-error']
 
 def test_emitter_error(error_filename, verbose=False):
     events = list(yaml.load(open(error_filename, 'rb'),
-                    Loader=test_emitter.EventsLoader))
+                    Loader=EventsLoader))
     try:
         yaml.emit(events)
-    except yaml.YAMLError, exc:
+    except yaml.YAMLError as exc:
         if verbose:
-            print "%s:" % exc.__class__.__name__, exc
+            print("%s:" % exc.__class__.__name__, exc)
     else:
         raise AssertionError("expected an exception")
 
@@ -51,17 +53,12 @@ def test_dumper_error(error_filename, verbose=False):
     code = open(error_filename, 'rb').read()
     try:
         import yaml
-        from StringIO import StringIO
-        exec code
-    except yaml.YAMLError, exc:
+        from io import StringIO
+        exec(code)
+    except yaml.YAMLError as exc:
         if verbose:
-            print "%s:" % exc.__class__.__name__, exc
+            print("%s:" % exc.__class__.__name__, exc)
     else:
         raise AssertionError("expected an exception")
 
 test_dumper_error.unittest = ['.dumper-error']
-
-if __name__ == '__main__':
-    import test_appliance
-    test_appliance.run(globals())
-
