+++ tests/test_reader.py
@@ -1,15 +1,15 @@
 
-import yaml.reader
 import codecs
+import yaml.reader
 
 def _run_reader(data, verbose):
     try:
         stream = yaml.reader.Reader(data)
-        while stream.peek() != u'\0':
+        while stream.peek() != '\0':
             stream.forward()
-    except yaml.reader.ReaderError, exc:
+    except yaml.reader.ReaderError as exc:
         if verbose:
-            print exc
+            print(exc)
     else:
         raise AssertionError("expected an exception")
 
@@ -18,7 +18,7 @@ def test_stream_error(error_filename, verbose=False):
     _run_reader(open(error_filename, 'rb').read(), verbose)
     for encoding in ['utf-8', 'utf-16-le', 'utf-16-be']:
         try:
-            data = unicode(open(error_filename, 'rb').read(), encoding)
+            data = open(error_filename, 'rb').read().decode(encoding)
             break
         except UnicodeDecodeError:
             pass
@@ -28,8 +28,3 @@ def test_stream_error(error_filename, verbose=False):
     _run_reader(codecs.open(error_filename, encoding=encoding), verbose)
 
 test_stream_error.unittest = ['.stream-error']
-
-if __name__ == '__main__':
-    import test_appliance
-    test_appliance.run(globals())
-
