+++ tests/test_structure.py
@@ -1,6 +1,8 @@
 
-import yaml, canonical
+import yaml
 import pprint
+
+from tests import canonical
 
 def _convert_structure(loader):
     if loader.check_event(yaml.ScalarEvent):
@@ -34,7 +36,7 @@ def _convert_structure(loader):
 
 def test_structure(data_filename, structure_filename, verbose=False):
     nodes1 = []
-    nodes2 = eval(open(structure_filename, 'rb').read())
+    nodes2 = eval(open(structure_filename, 'r').read())
     try:
         loader = yaml.Loader(open(data_filename, 'rb'))
         while loader.check_event():
@@ -48,24 +50,24 @@ def test_structure(data_filename, structure_filename, verbose=False):
         assert nodes1 == nodes2, (nodes1, nodes2)
     finally:
         if verbose:
-            print "NODES1:"
+            print("NODES1:")
             pprint.pprint(nodes1)
-            print "NODES2:"
+            print("NODES2:")
             pprint.pprint(nodes2)
 
 test_structure.unittest = ['.data', '.structure']
 
 def _compare_events(events1, events2, full=False):
-    assert len(events1) == len(events2), (len(events1), len(events2))
+    assert len(events1) == len(events2)
     for event1, event2 in zip(events1, events2):
-        assert event1.__class__ == event2.__class__, (event1, event2)
+        assert event1.__class__ == event2.__class__
         if isinstance(event1, yaml.AliasEvent) and full:
-            assert event1.anchor == event2.anchor, (event1, event2)
+            assert event1.anchor == event2.anchor
         if isinstance(event1, (yaml.ScalarEvent, yaml.CollectionStartEvent)):
-            if (event1.tag not in [None, u'!'] and event2.tag not in [None, u'!']) or full:
-                assert event1.tag == event2.tag, (event1, event2)
+            if (event1.tag not in [None, '!'] and event2.tag not in [None, '!']) or full:
+                assert event1.tag == event2.tag
         if isinstance(event1, yaml.ScalarEvent):
-            assert event1.value == event2.value, (event1, event2)
+            assert event1.value == event2.value
 
 def test_parser(data_filename, canonical_filename, verbose=False):
     events1 = None
@@ -76,9 +78,9 @@ def test_parser(data_filename, canonical_filename, verbose=False):
         _compare_events(events1, events2)
     finally:
         if verbose:
-            print "EVENTS1:"
+            print("EVENTS1:")
             pprint.pprint(events1)
-            print "EVENTS2:"
+            print("EVENTS2:")
             pprint.pprint(events2)
 
 test_parser.unittest = ['.data', '.canonical']
@@ -92,9 +94,9 @@ def test_parser_on_canonical(canonical_filename, verbose=False):
         _compare_events(events1, events2, full=True)
     finally:
         if verbose:
-            print "EVENTS1:"
+            print("EVENTS1:")
             pprint.pprint(events1)
-            print "EVENTS2:"
+            print("EVENTS2:")
             pprint.pprint(events2)
 
 test_parser_on_canonical.unittest = ['.canonical']
@@ -124,48 +126,48 @@ def test_composer(data_filename, canonical_filename, verbose=False):
             _compare_nodes(node1, node2)
     finally:
         if verbose:
-            print "NODES1:"
+            print("NODES1:")
             pprint.pprint(nodes1)
-            print "NODES2:"
+            print("NODES2:")
             pprint.pprint(nodes2)
 
 test_composer.unittest = ['.data', '.canonical']
 
 def _make_loader():
-    global MyLoader
-
     class MyLoader(yaml.Loader):
         def construct_sequence(self, node):
             return tuple(yaml.Loader.construct_sequence(self, node))
         def construct_mapping(self, node):
             pairs = self.construct_pairs(node)
-            pairs.sort()
+            pairs.sort(key=(lambda i: str(i)))
             return pairs
         def construct_undefined(self, node):
             return self.construct_scalar(node)
 
-    MyLoader.add_constructor(u'tag:yaml.org,2002:map', MyLoader.construct_mapping)
+    MyLoader.add_constructor('tag:yaml.org,2002:map', MyLoader.construct_mapping)
     MyLoader.add_constructor(None, MyLoader.construct_undefined)
 
-def _make_canonical_loader():
-    global MyCanonicalLoader
+    return MyLoader
 
+def _make_canonical_loader():
     class MyCanonicalLoader(yaml.CanonicalLoader):
         def construct_sequence(self, node):
             return tuple(yaml.CanonicalLoader.construct_sequence(self, node))
         def construct_mapping(self, node):
             pairs = self.construct_pairs(node)
-            pairs.sort()
+            pairs.sort(key=(lambda i: str(i)))
             return pairs
         def construct_undefined(self, node):
             return self.construct_scalar(node)
 
-    MyCanonicalLoader.add_constructor(u'tag:yaml.org,2002:map', MyCanonicalLoader.construct_mapping)
+    MyCanonicalLoader.add_constructor('tag:yaml.org,2002:map', MyCanonicalLoader.construct_mapping)
     MyCanonicalLoader.add_constructor(None, MyCanonicalLoader.construct_undefined)
 
+    return MyCanonicalLoader
+
 def test_constructor(data_filename, canonical_filename, verbose=False):
-    _make_loader()
-    _make_canonical_loader()
+    MyLoader = _make_loader()
+    MyCanonicalLoader = _make_canonical_loader()
     native1 = None
     native2 = None
     try:
@@ -174,14 +176,9 @@ def test_constructor(data_filename, canonical_filename, verbose=False):
         assert native1 == native2, (native1, native2)
     finally:
         if verbose:
-            print "NATIVE1:"
+            print("NATIVE1:")
             pprint.pprint(native1)
-            print "NATIVE2:"
+            print("NATIVE2:")
             pprint.pprint(native2)
 
 test_constructor.unittest = ['.data', '.canonical']
-
-if __name__ == '__main__':
-    import test_appliance
-    test_appliance.run(globals())
-
