+++ tox.ini
@@ -1,0 +1,13 @@
+[tox]
+envlist = py26, py27, py33, py34, py35, pypy, pypy3
+
+[testenv]
+deps =
+    pytest
+commands =
+    py.test --strict {posargs}
+
+[pytest]
+addopts = -r s
+markers =
+    skip: Skips a test, work around for pytest-dev/pytest#1472
